+++ CHANGELOG.rst
@@ -11,6 +11,10 @@ v33.0.0 (unreleased)
 - Remove the admin app and views.
   https://github.com/nexB/scancode.io/issues/645
 
+- Enhance the ``resolve_about_packages`` pipe to handle filename and checksum values.
+
+- Split the pipes unit tests into their own related submodule.
+
 v32.1.0 (2023-03-23)
 --------------------